+++ core/app/templates/index.html
