+++ core/devices/__init__.py
@@ -7,7 +7,7 @@ import cube_strip
 import lonely
 import big_cube_device
 import fft_device
-import app.app_device
+import app_device
 
 # Returns a list constructed devices given a list of initialisation data
 def construct_output_devices(output_devices_data_dict):
